refactor: create new branch
This new branch is now using vue-demi as the main development wrapper
to add support to vue 2 later.

Currently, this branch is experimenting 2 types of usage. First, the purely
component approach which has more template codes but is more straight forward
and easier to understand. The second one is but using a more leaner approach
using composables. With only exposing useStripe with two main methors initStripe
and initElements, you'll easily get the stripe instance and the elements instance
which promotes a more robust usage because you already have the access to the
instances inside the script compared to the component where they're only in the
slot scope variables.